backend/requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
spacy==3.6.1
thinc==8.1.10
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
chardet==5.2.0
gunicorn==21.2.0
fpdf==1.7.2